requirements.txt
fastapi
python-dotenv
requests
uvicorn

# para subir archivos a FastAPI
python-multipart
pypdf
python-docx
